% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brickDrivers.R
\name{brickDrivers}
\alias{brickDrivers}
\title{Stack `eDrivers` objects as `eDriversBrick`}
\usage{
brickDrivers(drivers)
}
\arguments{
\item{drivers}{a vector of characters indicating the queried drivers as keys
or file names. Consult list of drivers using the `fetchList()` function to
identify desired drivers and use `FileName` or `Key` column to identify proper
spelling for queries.}
}
\value{
an `eDriversBrick` object containing a list with:
 [[1]]: rasterbrick with drivers queried
 [[2]]: list containing metadata for each driver queried
 [[3]]: list containing sources for each driver queried
}
\description{
Uses a list of `eDrivers` objects to create a `eDriversBrick`, which is a `rasterbrick`
with additionnal attributes and data associated with `eDrivers` objects
}
\examples{
\donttest{
res1 <- brickDrivers(c('CoastalDevelopment','PositiveSST'))
res1
summary(res1)
plot(res1)
}
}
