% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchDrivers.R
\name{fetchDrivers}
\alias{fetchDrivers}
\title{Downloads and imports driver layers}
\usage{
fetchDrivers(drivers, output = NULL, import = TRUE, brick = FALSE)
}
\arguments{
\item{drivers}{a vector of characters indicating the queried drivers as keys
or file names. Consult list of drivers using the `fetchList()` function to
identify desired drivers and use `FileName` or `Key` column to identify proper
spelling for queries.}

\item{output}{path to disk location of downloaded drivers. Default is `null`,
which uses the current working directory.}

\item{import}{a logical. Should queried drivers be imported in R session as
`eDrivers` objects or `eDriversBrick` objects (see brick parameters).}

\item{brick}{a logical. Should a `eDriversBrick` object be returned, which
a single `RasterBrick` object containing all drivers queried. See
`brickDrivers()` for further details.}
}
\value{
Saves queried drivers and accompanying attributes (metadata and
sources) to disk and imports data in R session if import is `TRUE`.
if brick is `FALSE`, `eDrivers` objects are imported, one for each
driver queried, containing a list with:
 [[1]]: Spatial data as a `RasterLayer`
 [[2]]: Metadata as a `yaml`
 [[3]]: Sources as a `BibEntry`
if brick is `TRUE`, an `eDriversBrick` object is imported containing a list with:
 [[1]]: RasterBrick` with drivers queried
 [[2]]: list containing metadata for each driver queried
 [[3]]: list containing sources for each driver queried
}
\description{
Downloads and imports queried driver layers with accompanying attributes
(metadata and source) for a selection of drivers
}
\examples{
\dontrun{
# Example 1
res1 <- fetchDrivers(drivers = c("SST+", "SST-"), import = FALSE)

# Example 2
res2 <- fetchDrivers(drivers = c("SST+", "SST-"), import = TRUE, brick = FALSE)

# Example 3
res3 <- fetchDrivers(drivers = c("SST+", "SST-"), import = TRUE, brick = TRUE)
}
}
