% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetchList.R
\name{fetchList}
\alias{fetchList}
\title{Fetches the list of drivers available}
\usage{
fetchList()
}
\value{
An data.frame with details on drivers available containing driver
groups (Groups), driver names (Drivers), driver keys used for queries
(Key), name of files (FileName), which can also be used for
queries, and data source (Source).
}
\description{
Fetches and return a table with the list of driver available through the
`eDrivers` R package.
}
\examples{
\donttest{
res1 <- fetchList()
res1
}
}
