% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramDrivers.R
\name{paramDrivers}
\alias{paramDrivers}
\title{Sets required parameters for internal functions and checks for errors in
queried drivers}
\usage{
paramDrivers(drivers)
}
\arguments{
\item{drivers}{a vector of characters indicating the queried drivers as keys
or file names. Consult list of drivers using the `fetchList()` function to
identify desired drivers and use `FileName` or `Key` column to identify proper
spelling for queries.}
}
\value{
drNames object containing driver names to use internally in `eDrivers` functions
}
\description{
Sets required parameters for internal functions and checks for errors in
queried drivers
}
\examples{
\donttest{
paramDrivers()
}
}
