% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.eDrivers}
\alias{plot.eDrivers}
\alias{plot.eDriversBrick}
\title{Plot methods for objects of class `eDrivers*`}
\usage{
\method{plot}{eDrivers}(x, ...)

\method{plot}{eDriversBrick}(x, ...)
}
\arguments{
\item{x}{An object of class `eDrivers*`.}

\item{...}{ignored.}
}
\value{
The plot of the `eDrivers*` object.
}
\description{
Plots the outputs of an `eDrivers*` object call
}
