% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.eDrivers}
\alias{print.eDrivers}
\alias{print.eDriversBrick}
\title{Print methods for objects of class `eDrivers*`}
\usage{
\method{print}{eDrivers}(x, ...)

\method{print}{eDriversBrick}(x, ...)
}
\arguments{
\item{x}{object of class `eDrivers*`.}

\item{...}{ignored.}
}
\value{
The JSON result.
}
\description{
Print the outputs of an `eDrivers*` object call
}
